feat(RandomNumberGenerator): 랜덤 값 생성 책임을 분리한다
