+++ src/main/java/domain/numbergenerator/NumberGenerator.java
@@ -1,0 +1,5 @@
+package domain.numbergenerator;
+
+public interface NumberGenerator {
+    int generate();
+}
