+++ src/test/java/domain/LottoTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import domain.numbergenerator.NumberGenerator;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -10,6 +11,14 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 class LottoTest {
 
+    public static Stream<Arguments> getLottoNumbersInputData() {
+        return Stream.of(
+                Arguments.of(List.of(1, 2, 3, 10, 11, 12), 3, false),
+                Arguments.of(List.of(1, 2, 3, 7, 10, 11), 3, true),
+                Arguments.of(List.of(1, 2, 3, 4, 5, 6), 6, false)
+        );
+    }
+
     @MethodSource("getLottoNumbersInputData")
     @ParameterizedTest
     void 당첨_번호와_비교하여_일치_결과를_반환한다(
@@ -17,7 +26,7 @@ class LottoTest {
         List<Integer> winningNumbers = List.of(1, 2, 3, 4, 5, 6);
         int bonusNumber = 7;
 
-        Lotto lotto = new Lotto(lottoNumbers);
+        Lotto lotto = Lotto.issueByNumberGenerator(new FixedNumberGenerator(lottoNumbers));
 
         MatchDto matchDto = lotto.getMatchDto(winningNumbers, bonusNumber);
 
@@ -25,12 +34,21 @@ class LottoTest {
         assertThat(matchDto.hasBonusNumber()).isEqualTo(hasBonusNumber);
     }
 
-    public static Stream<Arguments> getLottoNumbersInputData() {
-        return Stream.of(
-                Arguments.of(List.of(1, 2, 3, 10, 11, 12), 3, false),
-                Arguments.of(List.of(1, 2, 3, 7, 10, 11), 3, true),
-                Arguments.of(List.of(1, 2, 3, 4, 5, 6), 6, false)
-        );
+    public static class FixedNumberGenerator implements NumberGenerator {
+        private final List<Integer> numbers;
+        private int index = 0;
+
+        public FixedNumberGenerator(List<Integer> numbers) {
+            this.numbers = numbers;
+        }
+
+        @Override
+        public int generate() {
+            int generatedNumber = numbers.get(index);
+            index += 1;
+
+            return generatedNumber;
+        }
     }
 
 }
